+++ tests/periph_spi/main.c
@@ -110,7 +110,7 @@ int cmd_init(int argc, char **argv)
 
     /* parse the given SPI device */
     dev = atoi(argv[1]);
-    if (dev < 0 || dev >= SPI_NUMOF) {
+    if (dev < 0 || dev >= (int)SPI_NUMOF) {
         puts("error: invalid SPI device specified");
         return 1;
     }
